+++ phaser_core/src/backend/correlation/base-spatial-correlation.cc
@@ -1,4 +1,4 @@
-#include "phaser/backend/alignment/base-spatial-correlation.h"
+#include "phaser/backend/correlation/base-spatial-correlation.h"
 
 namespace alignment {
